+++ src/views/dashboard/Widgets.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { db, selectWidgets, WidgetPosition, WidgetState } from "../../db/state";
+import { selectWidgets } from "../../db/select";
+import { db, WidgetPosition, WidgetState } from "../../db/state";
 import { useSelector, useValue } from "../../lib/db/react";
 import Slot from "./Slot";
 import "./Widgets.sass";
